+++ tools/nnf/pyproject.toml
@@ -7,7 +7,7 @@ name = "nnf"
 version = "0.1.0"
 requires-python = ">=3.6"
 dependencies = [
-    "kubernetes>=12.0,<27",
+    "kubernetes>=12.0,<24",
 ]
 
 [project.scripts]
